+++ src/js/pip.js
@@ -1,6 +1,0 @@
-window.addEventListener("load", () => {
-    window.addEventListener("keydown", (e) => {
-        if (e.key == "F4") e.preventDefault();
-        
-    });
-});